refactor: relocate rich text editor for reuse
