+++ src/components/RichTextEditor/index.ts
@@ -1,0 +1,3 @@
+import RichTextEditor from './RichTextEditor';
+
+export default RichTextEditor;
